Crate palmrs_database

Support for reading, and eventually writing, the Palm OS database formats (PRC and PDB)

Modules

header

The common file header used by both the PRC and PDB databases

info

App info & sort info parsers and helpers

record

Individual database record handling

time

PalmTimestamp type & conversion methods

Structs

PalmDatabase

A representation of a Palm OS database file

PdbDatabase

Implementation of DatabaseFormat for PDB databases

PdbWithCategoriesDatabase

Implementation of DatabaseFormat for PDB databases that contain category information

PrcDatabase

Implementation of DatabaseFormat for PRC databases

Traits

DatabaseFormat

Helper trait for database format types