Trait palmrs_database::info::ExtraInfoRecord

pub trait ExtraInfoRecord: Sized + Debug {
    fn from_bytes(
        hdr: &DatabaseHeader, 
        data: &[u8], 
        pos: usize
    ) -> Result<Self, Error>;

    fn write_bytes(&self) -> Result<Vec<u8>, Error>;

    fn data_empty(&self) -> bool;

    fn data_item_categories(&self) -> Option<Vec<ExtraInfoCategory>> { ... }
}

Helper trait for decoding & encoding “extra data” records (app info / sort info)

Required methods

fn from_bytes(
hdr: &DatabaseHeader,
data: &[u8],
pos: usize
) -> Result<Self, Error>

Read the record header from the given byte array

fn write_bytes(&self) -> Result<Vec<u8>, Error>

Write the record header to a new Vec<u8>

fn data_empty(&self) -> bool

Whether this ExtraInfoRecord contains no data

Provided methods

fn data_item_categories(&self) -> Option<Vec<ExtraInfoCategory>>

If the record contains a list of item categories, return those categories