Module palmrs_database::time

PalmTimestamp type & conversion methods

Palm OS has two epochs used for storing time - the “old Palm epoch” (seconds since 1904-01-01 00:00:00), and the standard UNIX epoch (seconds since 1970-01-01 00:00:00). If the “old Palm epoch” is used for storing the time, it is stored as an unsigned 32-bit integer, but if the UNIX epoch is used, it is stored as a signed 32-bit integer.

This module provides the PalmTimestamp helper type, which can be used within larger data structures to provide automatic timestamp format conversion. This module also provides various helper methods for timestamp format detection, and conversion between the timestamp formats.

Structs

PalmTimestamp

Type representing a Palm OS timestamp

Constants

SECONDS_BETWEEN_PALM_EPOCHS

The number of seconds between the two Palm OS timestamp epochs.

Functions

is_palm_epoch

Check if the given timestamp is using the “old Palm epoch”

palm_ts_to_unix_ts

Convert an “old Palm epoch” timestamp to a UNIX epoch timestamp

unix_ts_to_palm_ts

Convert a UNIX epoch timestamp to an “old Palm epoch” timestamp