Trait palmrs_database::info::ExtraInfoRecord

pub trait ExtraInfoRecord: Sized + Debug {
    const SIZE: usize;

    // Required methods
    fn from_bytes(
        hdr: &DatabaseHeader,
        data: &[u8],
        pos: usize
    ) -> Result<Self, Error>;
    fn to_bytes(&self) -> Result<Vec<u8>, Error>;
    fn data_empty(&self) -> bool;

    // Provided method
    fn data_item_categories(&self) -> Option<Vec<ExtraInfoCategory>> { ... }
}

Helper trait for decoding & encoding “extra data” records (app info / sort info)

Required Associated Constants

const SIZE: usize

Size in bytes (packed) which the ExtraInfoRecord occupies in the pdb/prc

Required Methods

fn from_bytes( hdr: &DatabaseHeader, data: &[u8], pos: usize ) -> Result<Self, Error>

Read the record header from the given byte array

fn to_bytes(&self) -> Result<Vec<u8>, Error>

Write the record header to a new Vec<u8>

fn data_empty(&self) -> bool

Whether this ExtraInfoRecord contains no data

Provided Methods

fn data_item_categories(&self) -> Option<Vec<ExtraInfoCategory>>

If the record contains a list of item categories, return those categories