Trait palmrs_database::record::DatabaseRecord

pub trait DatabaseRecord: Sized + Debug {
    fn struct_len() -> usize;

    fn from_bytes(data: &[u8], pos: usize) -> Result<Self, Error>;

    fn name_str(&self) -> Option<&str>;

    fn attributes(&self) -> Option<u32>;

    fn data_offset(&self) -> u32;

    fn data_len(&self) -> Option<u32>;
}

Helper trait for database record types

Required methods

fn struct_len() -> usize

The length of the record header, in bytes

fn from_bytes(data: &[u8], pos: usize) -> Result<Self, Error>

Read the record header from the given byte array.

fn name_str(&self) -> Option<&str>

Return the record’s name, if known

fn attributes(&self) -> Option<u32>

Return the record’s attributes, if known

fn data_offset(&self) -> u32

Return the offset, from the start of the database file, of the record’s data.

fn data_len(&self) -> Option<u32>

Return the length of the record’s data, if known.